fix(03-bojo Step 5): Long→Double + 카운터 5dfa203 패턴 + Target 추적 + multi-target count 메타
큰 데이터셋 (14897 IF) 검증 시점에서 노출된 da653f2 (3/9) 의 숨은 회귀 4건을 의도 정합 차원으로 복원.

[1] IF entity Long → Double (94e2127 Phase 5 회귀)
- IfRsvSecObsvdata.gwdep/gwtemp/ec, IfRsvSecJewon.pyogo/guldep/guldia/casingHeight (7 fields)
- 표준 = NUMBER (DDL/외부 PG/DMZ entity 모두 소수 허용)
- 6.27 → "6" 데이터 손실 차단

[2] writeCount 단위 5dfa203 패턴 복원 (da653f2 의 비공개 변경)
- writeCount = obsvSuccess (14897, IF 단위) → += inserted (44691, EAV 1:3 팽창)
- mappingName="obsvdata" 통일 (main path / fail path 정합)

[3] Target execution_id 추적 (행 클릭 데이터 표시)
- batchUpsertLink: MERGE INTO 에 execution_id 컬럼 추가 (Oracle + PG)
- ensureResultId: INSERT 에 execution_id 컬럼 추가 + 시그니처 변경

[4] sync_log multi-target count 메타 (per-table → per-mapping 전환 시 누락 메타 보강)
- target_tables JSON: ["str"] → [{"name":..,"count":N}]
- TableStatsDto.targetCounts: Map<String, Long>
- ExecutionDataController.parseTargetCounts 신규 메서드 (backward-compat)
- Loader: saveSyncLogMappingWithCounts 신규 helper (다른 1:1 Loader 영향 0)
- frontend: TableStats.targetCounts + flatTableStats fallback 룰

별 사이클 issue:
- VER-014 (Tibero NUMBER(p,s) DDL precision 명시 + BigDecimal 마이그레이션 검토)
- VER-015 (TM_GD970101 / TM_GD980002 추적 비대상 정책 확정 — CLOSE 후보)

검증 (사용자 화면): PM 44691 / TM_GD970101 3618 / TM_GD980002 1206 + 6.27 보존 + PM 펼침 추적 + TM_* 표시만

+++ infolink-agent-bojo-internal/src/main/java/com/infolink/agent/bojo/entity/iftable/IfRsvSecJewon.java
@@ -49,22 +49,22 @@ public class IfRsvSecJewon {
     private String y;
 
     @Column(name = "PYOGO")
-    private Long pyogo;
+    private Double pyogo;
 
     @Column(name = "INSDATE")
     private LocalDateTime insdate;
 
     @Column(name = "GULDEP")
-    private Long guldep;
+    private Double guldep;
 
     @Column(name = "GULDIA")
-    private Long guldia;
+    private Double guldia;
 
     @Column(name = "REGDATE")
     private LocalDateTime regdate;
 
     @Column(name = "CASING_HEIGHT")
-    private Long casingHeight;
+    private Double casingHeight;
 
     @Column(name = "SOURCE_REFS", length = 4000, nullable = false)
     private String sourceRefs;
